Rename haiku validator to better communicate usage
